refactor: extract shared routes into shared_bp Blueprint
Move /, /render, /renderPNG, /encode, /decode into shared/routes.py
with a dedicated shared_bp Blueprint. Register it in app.py alongside
the existing plantuml Blueprint.
